this commit will organize my components for ease of access
