+++ frontend/src/App.tsx
@@ -10,7 +10,7 @@ import NoMatchPage from "./pages/NoMatchPage";
 import GroceryList from "./pages/GroceryList";
 import SignupForm from "./pages/SignupForm";
 import AccountPage from "./pages/AccountPage";
-import WithAuth from "./components/WithAuth";
+import WithAuth from "./components/authentication/WithAuth";
 import RecipeDiscoveryPage from "./pages/RecipeDiscoveryPage";
 import FavoritedRecipes from "./pages/FavoritedRecipes";
