+++ frontend/src/components/authentication/AuthForm.tsx
@@ -1,13 +1,13 @@
 import React from "react";
-import { Intolerances, Diets } from "../utils/enum";
+import { Intolerances, Diets } from "../../utils/enum";
 import { useState } from "react";
 import RegistrationPreferenceButtons from "./RegistrationPreferenceButtons";
 import {
   PreferenceCategoryEnum,
   AuthenticationFieldEnum,
-} from "../utils/constants";
+} from "../../utils/constants";
 import { Button, Box, Input, FormControl, FormLabel } from "@mui/joy";
-import type { GPAuthFormDataTypes } from "../utils/types";
+import type { GPAuthFormDataTypes } from "../../utils/types";
 
 type GPAuthFormEventProps = {
   handleRegistrationSubmit?: ({
