+++ frontend/src/pages/GroceryList.tsx
@@ -2,14 +2,14 @@ import type {
   GPErrorMessageTypes,
   GPIngredientDataTypes,
 } from "../utils/types";
-import AppHeader from "../components/AppHeader";
+import AppHeader from "../components/utils/AppHeader";
 import GroceryListDepartment from "../components/GroceryListDepartment";
 import { useState, useEffect } from "react";
-import IngredientModal from "../components/IngredientModal";
+import IngredientModal from "../components/ingredients/IngredientModal";
 import { Button, Box, Typography } from "@mui/joy";
 import { GROCERY_MODAL } from "../utils/constants";
-import TitledListView from "../components/TitledListView";
-import ErrorState from "../components/ErrorState";
+import TitledListView from "../components/utils/TitledListView";
+import ErrorState from "../components/utils/ErrorState";
 import { fetchGroceryList } from "../utils/databaseHelpers";
 import { CenteredTitledListStyle } from "../utils/UIStyle";
 const databaseUrl = import.meta.env.VITE_DATABASE_URL;
