+++ frontend/src/pages/SignupForm.tsx
@@ -7,9 +7,9 @@ import type {
   GPErrorMessageTypes,
 } from "../utils/types";
 import { handleNewUser, validateUserToken } from "../utils/databaseHelpers";
-import AuthForm from "../components/AuthForm";
-import AppHeader from "../components/AppHeader";
-import ErrorState from "../components/ErrorState";
+import AuthForm from "../components/authentication/AuthForm";
+import AppHeader from "../components/utils/AppHeader";
+import ErrorState from "../components/utils/ErrorState";
 import { handleAuthInputChange } from "../utils/utils";
 import { Box, Card } from "@mui/joy";
 import { useUser } from "../contexts/UserContext";
